% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnps.R
\name{access_gnps}
\alias{access_gnps}
\title{Access GNPS information}
\usage{
access_gnps(gnpsID, ms2ldaID = NULL)
}
\arguments{
\item{gnpsID}{task id for MZmine workflow.}

\item{ms2ldaID}{task id for ms2lda workflow.}
}
\value{
list of dataframes with the results from workflows.
}
\description{
This function uses GNPS task id and optionally ms2lda task id to retrieve information from ProteoSAFe api.
}
\keyword{gnps}
