% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_heatmap.R
\name{format_heatmap}
\alias{format_heatmap}
\title{Wrapper to format heamap from defined GNPS-MZmine inputs}
\usage{
format_heatmap(featureTable, metadataTable, selectField = NULL,
  selectValue = NULL, factorColList = list(),
  colorScale = rev(heat.colors(75)), main = "", ...)
}
\arguments{
\item{featureTable}{MZmine format feature table.}

\item{metadataTable}{qiime2 format metadata table.}

\item{selectField}{metadata field to subselect. Default is NULL.}

\item{selectValue}{metadata field value to subselect. Default is NULL.}

\item{factorColList}{nested list containing lists of: a vector of colors, one for each factor level, and a string with a metadata field name. Default is list().}

\item{colorScale}{hexadecimal vector of colors. Default is heat.colors(75).}

\item{main}{title for heatmap plot. Default is ''.}
}
\value{
heatmap list.
}
\description{
This function allows you to subset tables, create multiple colored bars for samples, change color scales, etc.
}
\keyword{heatmap}
